beer-league-dashboard/backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Flask-RESTX==1.3.0
SQLAlchemy==2.0.23
pandas==2.1.4
python-dotenv==1.0.0
requests==2.31.0
